pyproject.toml
[project]
name = "training-app"
version = "0.0.1"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "asyncpg",
    "python-dotenv",
    "alembic",
    "pydantic[email]",
    "pydantic-settings",
    "pyjwt",
    "psycopg2-binary",
    "cryptography",
    "python-jose"
]
